Tried adding a scene/a tilemap
I'm experiencing problems with the tilemap. The tiles on the Tile Palette aren't showing up. I can't figure out how to change the size of the sprite that is drawn. The player doesn't appear to be able to jump more than once because it doesn't count coliding with the Tilemap collider as landing.